flask-app/requirements.txt
flask
google-cloud-storage
google-cloud-video-transcoder